+++ src/useCases/posts.use.js
@@ -1,10 +1,16 @@
 import { Post } from '../models/posts.model.js'
 import { Author } from '../models/authors.model.js'
-import { Reaction } from '../models/reactions.model.js'
 import { StatusHttp } from '../libs/statusHttp.js'
+import { Reaction } from '../models/reactions.model.js'
+
 
 function getAll () {
-  return Post.find({}).populate({ path: 'reactions' })
+  return Post.find({}).populate({
+    path: 'reactions',
+    populate: {
+      path: "author", select: ['name']
+    }
+  })
 }
 
 async function create (newPost, userCurrent) {
@@ -27,7 +33,7 @@ async function update (idPost, unupdatedPost, updatedAt= Date.now()) {
 async function getById (idPost) {
   const postFound = await Post.findById(idPost)
   if (!postFound) throw new StatusHttp('Post not found', 400)
-  return Post.findById(idPost).populate({ path: 'comments' }).populate({ path: 'reactions' })
+  return Post.findById(idPost)
 }
 
 async function deleteById (idPost) {
@@ -36,14 +42,7 @@ async function deleteById (idPost) {
   return Post.findByIdAndDelete(idPost)
 }
 
-async function addReaction (postId, userCurrent) {
-  console.log({ ...postId, author: userCurrent })
-  const reactionCreated = await Reaction.create({author: userCurrent })
-  await Post.findByIdAndUpdate(postId, {
-    $push: { reactions: reactionCreated._id }
-  })
-  return reactionCreated
-}
+
 
 export {
   getAll,
@@ -51,5 +50,4 @@ export {
   update,
   deleteById,
   getById,
-  addReaction
 }
